Добавляет сервис и контроллер для модуля user
