+++ taskforce/apps/users/src/app/auth/auth.http
@@ -4,16 +4,13 @@ Content-Type: application/json
 
 {
   "name": "Chip Hawk",
-  "email": "qwerty@gmail.com",
+  "email": "werty@gmail.com",
   "city": "Москва",
   "passwordHash": "123456",
   "role": "Исполнитель",
   "avatar": "silicon.jpg",
   "dateBirth": "2000-05-14"
 }
-
-### Find user
-GET http://localhost:3333/api/auth/1710be71-4832-47ec-b79e-8aca1964a06c HTTP/1.1
 
 ### Login user
 POST http://localhost:3333/api/auth/login HTTP/1.1
@@ -22,13 +19,4 @@ Content-Type: application/json
 {
   "email": "qwerty@gmail.com",
   "passwordHash": "123456"
-}
-
-### Update user password
-PATCH http://localhost:3333/api/auth/dc8ca73a-550f-4d19-8d56-c57b85c874e3 HTTP/1.1
-Content-Type: application/json
-
-{
-  "currentPassword": "123456",
-  "newPassword": "654321"
 }
