+++ taskforce/apps/users/src/app/auth/auth.service.ts
@@ -5,7 +5,6 @@ import { CreateUserDto } from './dto/create-user.dto';
 import { UserMemoryRepository } from '../user/user-memory.repository';
 import { UserEntity } from '../user/user.entity';
 import { LoginUserDto } from './dto/login-user.dto';
-import { UpdatePasswordDto } from './dto/update-password-user.dto';
 
 @Injectable()
 export class AuthService {
@@ -32,16 +31,6 @@ export class AuthService {
     return this.userMemoryRepository.create(userEntity);
   }
 
-  async findUser(id: string) {
-    const existUser = await this.userMemoryRepository.findById(id);
-
-    if (!existUser) {
-      throw new Error(`User with this id doesn't exist`)
-    }
-
-    return existUser;
-  }
-
   async verifyUser(dto: LoginUserDto) {
     const {email, passwordHash} = dto;
     const existUser = await this.userMemoryRepository.findByEmail(email);
@@ -57,26 +46,5 @@ export class AuthService {
     }
 
     return userEntity.toObject();
-  }
-
-  async updatePassword(id: string, dto: UpdatePasswordDto) {
-    const existUser = await this.userMemoryRepository.findById(id);
-
-    if (!existUser) {
-      throw new Error(`User with this id doesn't exist`)
-    }
-
-    const userEntity = new UserEntity(existUser);
-
-    const isCorrectCurrentPassword = await userEntity.comparePassword(dto.currentPassword);
-
-    if (isCorrectCurrentPassword) {
-      await userEntity.hashPassword(dto.newPassword);
-      await this.userMemoryRepository.update(id, userEntity)
-
-      return userEntity;
-    }
-
-    throw new Error('The current password doesn\'t correct');
   }
 }
