+++ taskforce/apps/users/src/app/user/user.module.ts
@@ -1,8 +1,11 @@
 import { Module } from '@nestjs/common';
 import { UserMemoryRepository } from './user-memory.repository';
+import { UserController } from './user.controller';
+import { UserService } from './user.service';
 
 @Module({
-  providers: [UserMemoryRepository],
-  exports: [UserMemoryRepository]
+  providers: [UserMemoryRepository, UserService],
+  exports: [UserMemoryRepository],
+  controllers: [UserController],
 })
 export class UserModule {}
